requirements.txt
Django==3.1.6
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.15.0
tzlocal==2.1